# Make e2e benchmarks more stable