--- a/scripts/ci/locustfile.py
+++ b/scripts/ci/locustfile.py
@@ -1,4 +1,4 @@
-from locust import HttpUser, TaskSet, task, between
+from locust import HttpUser, TaskSet, task, constant
 import csv
 import random
 from collections import defaultdict
@@ -71,4 +71,4 @@ class OSRMTasks(TaskSet):
 class OSRMUser(HttpUser):
     tasks = [OSRMTasks]
     # random wait time between requests to not load server for 100%
-    wait_time = between(0.05, 0.5)
+    wait_time = constant(0.05)