--- a/scripts/ci/locustfile.py
+++ b/scripts/ci/locustfile.py
@@ -70,5 +70,4 @@ class OSRMTasks(TaskSet):
 
 class OSRMUser(HttpUser):
     tasks = [OSRMTasks]
-    # random wait time between requests to not load server for 100%
     wait_time = constant(0.05)