--- a/scripts/ci/run_benchmarks.sh
+++ b/scripts/ci/run_benchmarks.sh
@@ -35,7 +35,7 @@ function run_benchmarks_for_folder {
 
             # wait for osrm-routed to start
             curl --retry-delay 3 --retry 10 --retry-all-errors "http://127.0.0.1:5000/route/v1/driving/13.388860,52.517037;13.385983,52.496891?steps=true"
-            locust -f $FOLDER/scripts/ci/locustfile.py \
+            locust -f $FOLDER/../pr/scripts/ci/locustfile.py \
                 --headless \
                 --processes -1 \
                 --users 10 \