--- a/scripts/ci/run_benchmarks.sh
+++ b/scripts/ci/run_benchmarks.sh
@@ -18,7 +18,7 @@ function run_benchmarks_for_folder {
 
     FOLDER=$1
     RESULTS_FOLDER=$2
-    LOCUSTFILE_FOLDER=$3
+    SCRIPTS_FOLDER=$3
 
     mkdir -p $RESULTS_FOLDER
 
@@ -48,20 +48,9 @@ function run_benchmarks_for_folder {
 
         # wait for osrm-routed to start
         curl --retry-delay 3 --retry 10 --retry-all-errors "http://127.0.0.1:5000/route/v1/driving/13.388860,52.517037;13.385983,52.496891?steps=true"
-        # locust -f $LOCUSTFILE_FOLDER/scripts/ci/locustfile.py \
-        #     --headless \
-        #     --processes -1 \
-        #     --users 1 \
-        #     --spawn-rate 1 \
-        #     --host http://localhost:5000 \
-        #     --run-time 1m \
-        #     --csv=locust_results_$ALGORITHM \
-        #     --loglevel ERROR
-
-        # python3 $FOLDER/scripts/ci/process_locust_benchmark_results.py locust_results_$ALGORITHM $ALGORITHM $RESULTS_FOLDER
 
         for METHOD in route table nearest trip match; do
-         python3 $LOCUSTFILE_FOLDER/scripts/ci/e2e_benchmark.py --host http://localhost:5000 --method $METHOD --num_requests 10000 > $RESULTS_FOLDER/e2e_${METHOD}_${ALGORITHM}.bench
+            python3 $SCRIPTS_FOLDER/scripts/ci/e2e_benchmark.py --host http://localhost:5000 --method $METHOD --num_requests 1000 > $RESULTS_FOLDER/e2e_${METHOD}_${ALGORITHM}.bench
         done
 
         kill -0 $OSRM_ROUTED_PID