I realised that locust is not right tool for my initial intention for these end-to-end tests: I just wanted to measure performance metrics for singular requests, but locust is designed to measure performance of services under load (roughly speaking to answer the question "how many clients my backend can serve simultaneously?"), so I decided to drop it and just write simple script which we run against osrm-routed: it seems to be much more stable (plus we have more control on how this thing works).










<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1088.42<br/>plain u32: 1075.41<br/>aliased double: 976.189<br/>plain double: 987.892 | aliased u32: 1092.68<br/>plain u32: 1103.32<br/>aliased double: 949.025<br/>plain double: 948.202 |
| e2e_match_ch | Total: 3020.1361179351807ms<br/>Min time: 2.4335384368896484ms<br/>Mean time: 23.054474182711303ms<br/>Median time: 16.14093780517578ms<br/>95th percentile: 74.6084451675415ms<br/>99th percentile: 90.31732082366939ms<br/>Max time: 101.44424438476562ms | Total: 3010.5719566345215ms<br/>Min time: 2.584218978881836ms<br/>Mean time: 22.981465317820774ms<br/>Median time: 16.167879104614258ms<br/>95th percentile: 73.68934154510498ms<br/>99th percentile: 89.740252494812ms<br/>Max time: 101.81188583374023ms |
| e2e_match_mld | Total: 2053.5194873809814ms<br/>Min time: 1.9218921661376953ms<br/>Mean time: 15.675721277717416ms<br/>Median time: 9.299278259277344ms<br/>95th percentile: 49.50845241546631ms<br/>99th percentile: 58.59007835388179ms<br/>Max time: 67.52562522888184ms | Total: 2024.8019695281982ms<br/>Min time: 2.1724700927734375ms<br/>Mean time: 15.45650358418472ms<br/>Median time: 8.729219436645508ms<br/>95th percentile: 49.01862144470215ms<br/>99th percentile: 58.644270896911536ms<br/>Max time: 67.55566596984863ms |
| e2e_nearest_ch | Total: 1370.525598526001ms<br/>Min time: 1.1610984802246094ms<br/>Mean time: 1.370525598526001ms<br/>Median time: 1.2726783752441406ms<br/>95th percentile: 1.774907112121582ms<br/>99th percentile: 1.948857307434082ms<br/>Max time: 2.2411346435546875ms | Total: 1378.0488967895508ms<br/>Min time: 1.1670589447021484ms<br/>Mean time: 1.3780488967895508ms<br/>Median time: 1.2871026992797852ms<br/>95th percentile: 1.8047332763671873ms<br/>99th percentile: 1.8625473976135254ms<br/>Max time: 2.084970474243164ms |
| e2e_nearest_mld | Total: 1390.9969329833984ms<br/>Min time: 1.1823177337646484ms<br/>Mean time: 1.3909969329833984ms<br/>Median time: 1.302957534790039ms<br/>95th percentile: 1.8167495727539062ms<br/>99th percentile: 1.887824535369873ms<br/>Max time: 1.9884109497070312ms | Total: 1382.8415870666504ms<br/>Min time: 1.1601448059082031ms<br/>Mean time: 1.3828415870666504ms<br/>Median time: 1.2890100479125977ms<br/>95th percentile: 1.8177270889282227ms<br/>99th percentile: 1.8766164779663086ms<br/>Max time: 1.9295215606689453ms |
| e2e_route_ch | Total: 3208.8401317596436ms<br/>Min time: 1.4195442199707031ms<br/>Mean time: 3.2088401317596436ms<br/>Median time: 3.249645233154297ms<br/>95th percentile: 4.2049407958984375ms<br/>99th percentile: 4.689064025878906ms<br/>Max time: 5.119800567626953ms | Total: 3254.8928260803223ms<br/>Min time: 1.4653205871582031ms<br/>Mean time: 3.2548928260803223ms<br/>Median time: 3.3015012741088867ms<br/>95th percentile: 4.264521598815918ms<br/>99th percentile: 4.686849117279053ms<br/>Max time: 5.497217178344727ms |
| e2e_route_mld | Total: 3860.1200580596924ms<br/>Min time: 1.3704299926757812ms<br/>Mean time: 3.8601200580596924ms<br/>Median time: 3.88491153717041ms<br/>95th percentile: 5.312168598175049ms<br/>99th percentile: 5.838961601257323ms<br/>Max time: 6.595134735107422ms | Total: 3864.3553256988525ms<br/>Min time: 1.3842582702636719ms<br/>Mean time: 3.8643553256988525ms<br/>Median time: 3.9125680923461914ms<br/>95th percentile: 5.32221794128418ms<br/>99th percentile: 5.737259387969971ms<br/>Max time: 6.646871566772461ms |
| e2e_table_ch | Total: 17156.60524368286ms<br/>Min time: 2.0542144775390625ms<br/>Mean time: 17.15660524368286ms<br/>Median time: 16.37256145477295ms<br/>95th percentile: 31.3234806060791ms<br/>99th percentile: 33.16184997558594ms<br/>Max time: 43.32256317138672ms | Total: 17195.815324783325ms<br/>Min time: 2.2089481353759766ms<br/>Mean time: 17.195815324783325ms<br/>Median time: 16.62766933441162ms<br/>95th percentile: 31.308233737945553ms<br/>99th percentile: 33.37952375411987ms<br/>Max time: 34.783124923706055ms |
| e2e_table_mld | Total: 65905.26032447815ms<br/>Min time: 4.603385925292969ms<br/>Mean time: 65.90526032447815ms<br/>Median time: 62.540650367736816ms<br/>95th percentile: 127.98451185226439ms<br/>99th percentile: 135.07933139801025ms<br/>Max time: 167.42205619812012ms | Total: 65389.28771018982ms<br/>Min time: 4.3201446533203125ms<br/>Mean time: 65.38928771018982ms<br/>Median time: 62.259793281555176ms<br/>95th percentile: 126.63334608078003ms<br/>99th percentile: 135.02937078475952ms<br/>Max time: 141.98589324951172ms |
| e2e_trip_ch | Total: 11596.420049667358ms<br/>Min time: 1.9011497497558594ms<br/>Mean time: 11.596420049667358ms<br/>Median time: 11.13903522491455ms<br/>95th percentile: 19.943320751190186ms<br/>99th percentile: 21.72452688217163ms<br/>Max time: 25.699138641357422ms | Total: 11723.994970321655ms<br/>Min time: 1.718759536743164ms<br/>Mean time: 11.723994970321655ms<br/>Median time: 11.216282844543457ms<br/>95th percentile: 19.947946071624756ms<br/>99th percentile: 22.086760997772217ms<br/>Max time: 28.75804901123047ms |
| e2e_trip_mld | Total: 18968.85108947754ms<br/>Min time: 1.9826889038085938ms<br/>Mean time: 18.96885108947754ms<br/>Median time: 18.50736141204834ms<br/>95th percentile: 30.307972431182858ms<br/>99th percentile: 33.10990333557129ms<br/>Max time: 42.36292839050293ms | Total: 18795.0439453125ms<br/>Min time: 1.8734931945800781ms<br/>Mean time: 18.7950439453125ms<br/>Median time: 18.346309661865234ms<br/>95th percentile: 30.144953727722168ms<br/>99th percentile: 32.2248649597168ms<br/>Max time: 36.01717948913574ms |
| json-render | String: 6.57991ms<br/>Stringstream: 9.31487ms<br/>Vector: 6.91397ms | String: 6.55007ms<br/>Stringstream: 9.33229ms<br/>Vector: 6.94991ms |
| match_ch | Default radius: <br/>4.44212ms/req at 82 coordinate<br/>0.0541722ms/coordinate<br/>Radius 5m: <br/>4.4251ms/req at 82 coordinate<br/>0.0539646ms/coordinate<br/>Radius 10m: <br/>15.4078ms/req at 82 coordinate<br/>0.1879ms/coordinate<br/>Radius 15m: <br/>36.5705ms/req at 82 coordinate<br/>0.445981ms/coordinate<br/>Radius 30m: <br/>313.592ms/req at 82 coordinate<br/>3.82429ms/coordinate | Default radius: <br/>4.42772ms/req at 82 coordinate<br/>0.0539966ms/coordinate<br/>Radius 5m: <br/>4.42033ms/req at 82 coordinate<br/>0.0539065ms/coordinate<br/>Radius 10m: <br/>15.1491ms/req at 82 coordinate<br/>0.184745ms/coordinate<br/>Radius 15m: <br/>36.8513ms/req at 82 coordinate<br/>0.449406ms/coordinate<br/>Radius 30m: <br/>313.921ms/req at 82 coordinate<br/>3.82831ms/coordinate |
| match_mld | Default radius: <br/>2.7886ms/req at 82 coordinate<br/>0.0340074ms/coordinate<br/>Radius 5m: <br/>2.7461ms/req at 82 coordinate<br/>0.033489ms/coordinate<br/>Radius 10m: <br/>10.1172ms/req at 82 coordinate<br/>0.123381ms/coordinate<br/>Radius 15m: <br/>25.8181ms/req at 82 coordinate<br/>0.314855ms/coordinate<br/>Radius 30m: <br/>303.603ms/req at 82 coordinate<br/>3.70248ms/coordinate | Default radius: <br/>2.79265ms/req at 82 coordinate<br/>0.0340567ms/coordinate<br/>Radius 5m: <br/>2.75403ms/req at 82 coordinate<br/>0.0335857ms/coordinate<br/>Radius 10m: <br/>10.2031ms/req at 82 coordinate<br/>0.124428ms/coordinate<br/>Radius 15m: <br/>25.8187ms/req at 82 coordinate<br/>0.314862ms/coordinate<br/>Radius 30m: <br/>302.438ms/req at 82 coordinate<br/>3.68827ms/coordinate |
| osrm_contract | Time: 95.62s Peak RAM: 185.58MB | Time: 95.64s Peak RAM: 187.73MB |
| osrm_customize | Time: 1.33s Peak RAM: 115.12MB | Time: 1.34s Peak RAM: 114.95MB |
| osrm_extract | Time: 12.22s Peak RAM: 412.05MB | Time: 12.38s Peak RAM: 407.05MB |
| osrm_partition | Time: 2.30s Peak RAM: 148.50MB | Time: 2.27s Peak RAM: 148.48MB |
| packedvector | random write:<br/>std::vector 9864.1 ms<br/>util::packed_vector 83305.1 ms<br/>slowdown: 8.44528<br/>random read:<br/>std::vector 8560.88 ms<br/>util::packed_vector 34049.8 ms<br/>slowdown: 3.97737 | random write:<br/>std::vector 9803.85 ms<br/>util::packed_vector 83247.4 ms<br/>slowdown: 8.49129<br/>random read:<br/>std::vector 8523.32 ms<br/>util::packed_vector 34015.3 ms<br/>slowdown: 3.99085 |
| route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>512.896ms<br/>0.512896ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>353.193ms<br/>0.353193ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>629.824ms<br/>0.629824ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>151.888ms<br/>0.151888ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>97.251ms<br/>0.097251ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>133.286ms<br/>0.133286ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>151.855ms<br/>0.151855ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>97.0303ms<br/>0.0970303ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>139.854ms<br/>0.139854ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>511.484ms<br/>0.511484ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>351.783ms<br/>0.351783ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>629.865ms<br/>0.629865ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>151.3ms<br/>0.1513ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>96.4286ms<br/>0.0964286ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>132.22ms<br/>0.13222ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>162.144ms<br/>0.162144ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>96.4002ms<br/>0.0964002ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>131.251ms<br/>0.131251ms/req |
| route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>643.418ms<br/>0.643418ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>439.773ms<br/>0.439773ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>817.063ms<br/>0.817063ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>266.84ms<br/>0.26684ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>160.953ms<br/>0.160953ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>287.715ms<br/>0.287715ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>262.433ms<br/>0.262433ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>160.7ms<br/>0.1607ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>288.713ms<br/>0.288713ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>642.086ms<br/>0.642086ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>454.098ms<br/>0.454098ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>832.13ms<br/>0.83213ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>268.607ms<br/>0.268607ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>159.485ms<br/>0.159485ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>285.314ms<br/>0.285314ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>255.545ms<br/>0.255545ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>159.86ms<br/>0.15986ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>287.728ms<br/>0.287728ms/req |
| rtree | 1 result:<br/>208.809ms ->  0.0208809 ms/query<br/>10 results:<br/>244.473ms ->  0.0244473 ms/query | 1 result:<br/>206.455ms ->  0.0206455 ms/query<br/>10 results:<br/>242.095ms ->  0.0242095 ms/query |
</details>
<!-- BENCHMARK_RESULTS_END -->